requirements.txt
langchain==0.1.7
crewai==0.11.2
pillow==10.2.0
matplotlib==3.8.3
diffusers==0.26.3
torch==2.1.2
huggingface-hub==0.20.3
tavily-python==0.3.1
streamlit==1.31.1
streamlit-drawable-canvas==0.9.3
